Refactor awe configuration properties

+++ awe-framework/awe-modules/awe-scheduler/src/main/java/com/almis/awe/scheduler/service/ExecutionService.java
@@ -44,11 +44,11 @@ public class ExecutionService {
   /**
    * Start timeout job
    *
-   * @param execution
-   * @param timeout
-   * @param process
+   * @param execution Task execution
+   * @param timeout Timeout task
+   * @param process Future process
    */
-  public void startTimeoutJob(TaskExecution execution, Integer timeout, Future process) {
+  public void startTimeoutJob(TaskExecution execution, long timeout, Future process) {
     JobDataMap data = new JobDataMap();
     data.put("id", execution.getKey());
     data.put(TASK_JOB_TIMEOUT, timeout);
@@ -66,8 +66,8 @@ public class ExecutionService {
   /**
    * Start progress job
    *
-   * @param execution
-   * @param averageTime
+   * @param execution Task execution
+   * @param averageTime Average time
    */
   public void startProgressJob(TaskExecution execution, Integer averageTime) {
 
@@ -88,7 +88,7 @@ public class ExecutionService {
   /**
    * Interrupt timeout timer
    *
-   * @param execution
+   * @param execution Task execution
    */
   public void interruptExecutionJobs(TaskExecution execution) {
     JobKey timeoutJobKey = getJobKey(execution, "TIMEOUT");
@@ -105,8 +105,8 @@ public class ExecutionService {
   /**
    * Start report job
    *
-   * @param task
-   * @param execution
+   * @param task Scheduler task
+   * @param execution Task execution
    */
   public void startReportJob(Task task, TaskExecution execution) {
     // If report send status doesn't match to execution status, exit
